+++ cli/commands/check/show.go
@@ -95,8 +95,8 @@ func printCheckToList(r *types.CheckConfig, writer io.Writer) {
 				Value: strconv.FormatBool(r.Stdin),
 			},
 			{
-				Label: "Source",
-				Value: r.Source,
+				Label: "Proxy Entity ID",
+				Value: r.ProxyEntityID,
 			},
 			{
 				Label: "Organization",